#1 - 코드 리팩토링
- 패키지 분리
> 기존 한 패키지에 있던 클래스들을 각 역할에 맞는 패키지로 분리
- 상품 메뉴 생성 변경
> 기존에 KioskApp.class에서 상품메뉴 리스트를 메뉴에 추가하였는데,
   메뉴가 생성되는 시점에 메뉴 클래스에서 상품 list를 생성하는게 맞는 거같아 Menu 클래스에서 생성하는 것으로 변경
