+++ src/com/sparta/study/task01/app/KioskApp.java
@@ -1,8 +1,12 @@
-package com.sparta.study.task01;
+package com.sparta.study.task01.app;
+
+import com.sparta.study.task01.constant.MenuCategory;
+import com.sparta.study.task01.exception.MenuNotFoundException;
+import com.sparta.study.task01.model.Menu;
+import com.sparta.study.task01.model.Order;
+import com.sparta.study.task01.model.Product;
 
 import java.util.*;
-
-import static com.sparta.study.task01.MenuCategory.*;
 
 public class KioskApp {
     private final List<Menu> menus = new ArrayList<>();
@@ -12,7 +16,7 @@ public class KioskApp {
         init();
     }
 
-    boolean start() throws Exception {
+    public boolean start() throws Exception {
         Scanner sc = new Scanner(System.in);
         System.out.println(logo);
         showMenu();
@@ -83,46 +87,16 @@ public class KioskApp {
         }
     }
 
-    private void addProducts(Menu menu) {
-        MenuCategory category = MenuCategory.valueOf(menu.getName());
-
-        switch (category) {
-            case COFFEE -> menu.setProducts(
-                    List.of(
-                            new Product("Sinchon Coffee", "맛있는 신촌커피", 6.1),
-                            new Product("Americano", "맛있는 아메리카노", 4.5),
-                            new Product("Coffee Latte", "맛있는 카페라떼", 5.0)
-                    )
-            );
-            case DRINK -> menu.setProducts(
-                    List.of(
-                            new Product("GrapeFruit Ade", "맛있는 자몽에이드", 5.5),
-                            new Product("Shine Muscat Green Grape Ade", "맛있는 샤인머스캣청포도에이드", 6.3),
-                            new Product("Jeju Matcha Frappe", "맛있는 제주말차프라페", 5.8)
-                    )
-            );
-            case CAKE -> menu.setProducts(
-                    List.of(
-                            new Product("Strawberry Chocolate Fresh Cream Cake", "한가득 올린 상큼한 딸기, 크런치 초코볼이 초콜릿 생크림 사이사이 씹히는 투썸 시그니처 케이크", 35.0),
-                            new Product("Mascarpone Tiramisu", "마스카포네 치즈 크림과 에스프레소 커피 시럽이 촉촉하게 어우러진 이탈리안 무스 케이크", 36.0),
-                            new Product("New York Cheese Cake", "부드럽고 촉촉한 식감과 진한 치즈의 맛을 느낄 수 있는 정통 뉴욕 스타일의 구움 치즈 케이크", 35.0)
-                    )
-            );
-        }
-    }
-
-
     private void init() {
         //메뉴 등록
         Arrays.stream(MenuCategory.values())
-                .forEach(menuCategory -> menus.add(
-                        new Menu(menuCategory.name(), menuCategory.getDescription())
-                ));
-
-        //상품 메뉴 등록
-        menus.stream()
-                .forEach(menu ->
-                        addProducts(menu));
+                .forEach(menuCategory -> {
+                            Menu menu = new Menu(menuCategory.name(), menuCategory.getDescription());
+                            //메뉴 카테고리별 상품 등록
+                            menu.addProducts(menuCategory);
+                            menus.add(menu);
+                        }
+                );
     }
 
     public static final String logo =
